p2/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
pydantic==2.5.3
opencv-python==4.9.0.80
mediapipe==0.10.9
openai-whisper==20231117
librosa==0.10.1
numpy==1.24.3
ffmpeg-python==0.2.0
websockets==12.0
